+++ packages/expo-modules-autolinking/CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### 💡 Others
 
+- Add experimental `tvos` and `macos` resolution ([#46344](https://github.com/expo/expo/pull/46344) by [@kitten](https://github.com/kitten))
+
 ## 56.0.13 — 2026-05-26
 
 ### 🐛 Bug fixes